Adding Disqus Comments and "Edit on Github" link at the end of each tutorial
Now each tutorial can have a Disqus comment section at the end of the
page! Only the <forum shortname> variable has to added when the astropy
developers register their site on Disqus

At the end of each tutorial , an "Edit on Github"button takes the user
to the Github directory of that particular tutorial

prepare_deploy modification simply allows us to have comments for each
page separated based on the name of the page and point to respective
Github page of the tutorials.