Add repository-dispatch workflow to deployment
This PR adds a "repository dispatch" step to the build workflow. What
this does is trigger a GitHub Actions run in the astropy/learn-astropy
repository. The payload for the trigger includes a reference to the
workflow artifact, which will allow the learn-astropy build to download
the tutorials artifact.

This requires a GitHub Secret to be set up on the tutorials repo named
`DISPATCH_GITHUB_TOKEN` that has a write access to the
astropy/learn-astropy repository.
See https://github.com/peter-evans/repository-dispatch#token